+++ packages/api/server/services/plan/create.js
@@ -1,0 +1,378 @@
+const ServiceError = require('#server/errors/ServiceError');
+const sequelize = require('#db/sequelize');
+const planModel = require('#server/models/planModel');
+const shantytownModel = require('#server/models/shantytownModel');
+const departementModel = require('#server/models/departementModel');
+const financeTypeModel = require('#server/models/financeTypeModel');
+const planManagerModel = require('#server/models/planManagerModel');
+const topicModel = require('#server/models/topicModel');
+const userModel = require('#server/models/userModel');
+const { addAttachments } = require('#server/models/permissionModel');
+const sanitize = require('#server/controllers/planController/_common/sanitize');
+
+module.exports = async (data, user) => {
+    // sanitize data
+    const planData = Object.assign({}, sanitize(data), {
+        createdBy: user.id,
+    });
+
+    // validate data
+    const errors = {};
+    function addError(field, error) {
+        if (errors[field] === undefined) {
+            errors[field] = [];
+        }
+
+        errors[field].push(error);
+    }
+
+    try {
+        const departement = await departementModel.findOne(planData.departement);
+        if (departement === null) {
+            addError('departement', 'Le département d\'intervention est obligatoire');
+        }
+    } catch (error) {
+        throw new ServiceError('departement_read_error', error);
+    }
+
+    let financeTypes;
+    try {
+        financeTypes = (await financeTypeModel.findAll()).reduce((acc, type) => Object.assign({}, acc, {
+            [type.uid]: type,
+        }), {});
+    } catch (error) {
+        throw new ServiceError('financeType_read_error', error);
+    }
+
+    let topics;
+    try {
+        topics = (await topicModel.findAll()).reduce((acc, topic) => Object.assign({}, acc, {
+            [topic.uid]: topic,
+        }), {});
+    } catch (error) {
+        throw new ServiceError('topic_read_error', error);
+    }
+
+    if (!planData.name) {
+        addError('name', 'Le nom de l\'action est obligatoire');
+    }
+
+    if (!planData.startedAt) {
+        addError('startedAt', 'La date de début de l\'action est obligatoire');
+    }
+
+    if (planData.expectedToEndAt && planData.expectedToEndAt <= planData.startedAt) {
+        addError('expectedToEndAt', 'La date de fin de l\'action ne peut pas être antérieure à la date de début');
+    }
+
+    if (typeof planData.inAndOut !== 'boolean') {
+        addError('in_and_out', 'Ce champ est obligatoire');
+    }
+
+    if (planData.topics.length === 0) {
+        addError('topics', 'Vous devez sélectionner au moins une thématique');
+    } else {
+        const unknownTopics = planData.topics.filter(uid => topics[uid] === undefined);
+        if (unknownTopics.length > 0) {
+            addError('topics', `Les thématiques suivantes n'ont pas été reconnues : ${unknownTopics.join(', ')}`);
+        }
+    }
+
+    if (!planData.goals) {
+        addError('goals', 'Les objectifs sont obligatoires');
+    }
+
+    if (!planData.locationType) {
+        addError('locationType', 'Le lieu est obligatoire');
+    } else if (['shantytowns', 'location', 'housing', 'other'].indexOf(planData.locationType) === -1) {
+        addError('locationType', 'Le type de lieu sélectionné n\'est pas reconnu');
+    }
+
+    switch (planData.locationType) {
+        case 'shantytowns':
+            if (planData.locationShantytowns.length === 0) {
+                addError('locationShantytowns', 'Vous devez sélectionner au moins un site');
+            } else {
+                try {
+                    const ids = await shantytownModel.findAll(user, [
+                        {
+                            shantytown_id: planData.locationShantytowns,
+                        },
+                    ]);
+
+                    if (ids.length !== planData.locationShantytowns.length) {
+                        addError('locationShantytowns', 'Un ou plusieurs sites sélectionnés n\'ont pas été retrouvés en base de données');
+                    }
+                } catch (error) {
+                    addError('locationShantytowns', 'Une erreur est survenue lors de la validation des sites');
+                }
+            }
+            break;
+
+        case 'location':
+            if (!planData.locationAddress) {
+                addError('locationAddress', 'L\'adresse du terrain est obligatoire');
+            }
+            break;
+
+        case 'other':
+            if (!planData.locationDetails) {
+                addError('locationDetails', 'Vous devez préciser les lieux de déroulement de l\'action');
+            }
+            break;
+
+        default:
+    }
+
+    if (!planData.government || planData.government.length === 0) {
+        addError('government', 'Vous devez désigner au moins une personne en charge du pilotage de l\'action');
+    } else {
+        try {
+            const users = await userModel.findByIds(user, planData.government);
+            if (users.length !== planData.government) {
+                addError('government', 'Une des personnes désignées comme pilote de l\'action n\'a pas été retrouvée en base de données');
+            } else if (
+                users.filter(manager => manager.organization.category.uid !== 'public_establishment').length > 0) {
+                addError('government', 'Au moins un des pilotes de l\'action ne fait partie d\'un service de l\'état');
+            }
+        } catch (error) {
+            addError('government', 'Une erreur est survenue lors de la validation du pilote de l\'action');
+        }
+    }
+
+    let associationContact;
+    if (!planData.associationContact) {
+        addError('association', 'Vous devez désigner la personne référente au sein de l\'opérateur ou service en charge de l\'intervention');
+    } else {
+        try {
+            associationContact = await userModel.findOne(planData.associationContact);
+            if (associationContact === null) {
+                addError('contact', 'La personne référente n\'a pas été retrouvée en base de données');
+            } else if (associationContact.organization.category.uid !== 'association') {
+                addError('association', 'Le service en charge de l\'intervention doit être une association');
+            }
+        } catch (error) {
+            addError('contact', 'Une erreur est survenue lors de la validation du service en charge de l\'intervention');
+        }
+    }
+
+    if (planData.finances) {
+        planData.finances.forEach(({ year, financeData }) => {
+            if (year > (new Date()).getFullYear()) {
+                addError('finances', `Il est impossible de saisir les financements pour l'année ${year}`);
+            } else {
+                financeData.forEach(({ amount, type }, index) => {
+                    const typeName = financeTypes[type] ? `'${financeTypes[type].name}'` : `de la ligne n°${index + 1}`;
+
+                    if (!type) {
+                        addError('finances', `Année ${year} : merci de préciser le type de financement pour la ligne n°${index + 1}`);
+                    } else if (financeTypes[type] === undefined) {
+                        addError('finances', `Année ${year} : le type de financement de la ligne n°${index + 1} n'est pas reconnu`);
+                    }
+
+                    if (Number.isNaN(amount)) {
+                        addError('finances', `Année ${year} : le montant du financement ${typeName} est invalide`);
+                    } else if (amount <= 0) {
+                        addError('finances', `Année ${year} : le montant du financement ${typeName} ne peut pas être négatif ou nul`);
+                    }
+                });
+            }
+        });
+    }
+
+    if (Object.keys(errors).length > 0) {
+        throw errors;
+    }
+
+    // insert into database
+    let finalPlanId;
+    try {
+        [finalPlanId] = await sequelize.transaction(async (t) => {
+            let locationId = null;
+            if (planData.locationType === 'location') {
+                const response = await sequelize.query(
+                    `INSERT INTO locations(
+                        address,
+                        latitude,
+                        longitude,
+                        created_by
+                    ) VALUES(
+                        :address,
+                        :latitude,
+                        :longitude,
+                        :createdBy
+                    ) RETURNING location_id AS id`,
+                    {
+                        replacements: Object.assign({}, planData.locationAddress, {
+                            createdBy: user.id,
+                        }),
+                        transaction: t,
+                    },
+                );
+
+                locationId = response[0][0].id;
+            }
+
+            const response = await sequelize.query(
+                `INSERT INTO plans2(
+                    name,
+                    started_at,
+                    expected_to_end_at,
+                    in_and_out,
+                    goals,
+                    fk_category,
+                    location_type,
+                    location_details,
+                    fk_location,
+                    created_by
+                ) VALUES(
+                    :name,
+                    :startedAt,
+                    :expectedToEndAt,
+                    :inAndOut,
+                    :goals,
+                    'autre',
+                    :locationType,
+                    :locationDetails,
+                    :fk_location,
+                    :createdBy
+                ) RETURNING plan_id AS id`,
+                {
+                    replacements: Object.assign({}, planData, {
+                        fk_location: locationId,
+                    }),
+                    transaction: t,
+                },
+            );
+            const planId = response[0][0].id;
+
+            await sequelize.query(
+                `INSERT INTO plan_departements(fk_plan, fk_departement, created_by)
+                VALUES (:planId, :departement, :createdBy)`,
+                {
+                    replacements: {
+                        planId,
+                        departement: planData.departement,
+                        createdBy: user.id,
+                    },
+                    transaction: t,
+                },
+            );
+
+            await sequelize.query(
+                `INSERT INTO plan_topics(fk_plan, fk_topic, created_by)
+                VALUES ${planData.topics.map(() => '(?, ?, ?)').join(', ')}`,
+                {
+                    replacements: planData.topics.reduce((acc, uid) => [
+                        ...acc,
+                        planId,
+                        uid,
+                        user.id,
+                    ], []),
+                    transaction: t,
+                },
+            );
+
+            if (planData.locationType === 'shantytowns') {
+                await sequelize.query(
+                    `INSERT INTO plan_shantytowns(fk_plan, fk_shantytown, created_by)
+                    VALUES ${planData.locationShantytowns.map(() => '(?, ?, ?)').join(', ')}`,
+                    {
+                        replacements: planData.locationShantytowns.reduce((acc, id) => [
+                            ...acc,
+                            planId,
+                            id,
+                            user.id,
+                        ], []),
+                        transaction: t,
+                    },
+                );
+            }
+
+            // insert into finances
+            const financeIds = await Promise.all(
+                planData.finances.map(({ year }) => sequelize.query(
+                    'INSERT INTO finances(fk_plan, year, created_by) VALUES (:planId, :year, :createdBy) RETURNING finance_id AS id',
+                    {
+                        replacements: {
+                            planId,
+                            year,
+                            createdBy: user.id,
+                        },
+                        transaction: t,
+                    },
+                )),
+            );
+
+            // insert into finance_rows
+            await Promise.all(
+                planData.finances.reduce((acc, { data: financeData }, index) => [
+                    ...acc,
+                    ...financeData.map(({
+                        amount, realAmount, type, details,
+                    }) => sequelize.query(
+                        `INSERT INTO finance_rows(fk_finance, fk_finance_type, amount, real_amount, comments, created_by)
+                        VALUES (:financeId, :type, :amount, :realAmount, :comments, :createdBy)`,
+                        {
+                            replacements: {
+                                financeId: financeIds[index][0][0].id,
+                                type,
+                                amount,
+                                realAmount,
+                                comments: details,
+                                createdBy: user.id,
+                            },
+                            transaction: t,
+                        },
+                    )),
+                ], []),
+            );
+
+            return Promise.all([
+                Promise.resolve(planId),
+
+                planManagerModel.create(planId, planData.government, user.id, t),
+
+                sequelize.query(
+                    `INSERT INTO plan_operators(fk_plan, fk_user, created_by)
+                    VALUES (:planId, :userId, :createdBy)`,
+                    {
+                        replacements: {
+                            planId,
+                            userId: planData.associationContact,
+                            createdBy: user.id,
+                        },
+                        transaction: t,
+                    },
+                ),
+
+                // pour le manager (utilisateur), on octroie les droits suivants sur l'action :
+                // list, read, update, et close
+                ...['list', 'read', 'update', 'close'].map(
+                    feature => planData.government.map(managerId => addAttachments([{ type: 'plan', id: planId }])
+                        .toUser(managerId)
+                        .onFeature(feature, 'plan', t)),
+                ),
+
+                // pour l'opérateur (structure), on octroie les droits suivants sur l'action :
+                // list, read, updateMarks
+                ...['list', 'read', 'updateMarks'].map(
+                    feature => addAttachments([{ type: 'plan', id: planId }])
+                        .toOrganization(associationContact.organization.id)
+                        .onFeature(feature, 'plan', t),
+                ),
+            ]);
+        });
+    } catch (error) {
+        throw new ServiceError('insert_error', error);
+    }
+
+    let plan = null;
+    try {
+        plan = await planModel.findOne(user, finalPlanId);
+    } catch (error) {
+        // ignore
+    }
+
+    return plan;
+};
